Setup state and event handling function in Main View Model

+++ finance_app/app/src/main/java/com/gkonstantakis/finance_app/data/database/entities/ProfitInfo.kt
@@ -11,5 +11,5 @@ data class ProfitInfo(
     var id: Int = 0,
 
     @ColumnInfo(name = "profit")
-    var profit: Double = 0.0,
+    var profit: Int = 0,
 )
