+++ finance_app/app/src/main/java/com/gkonstantakis/finance_app/data/models/ProfitInfoDomain.kt
@@ -2,6 +2,6 @@ package com.gkonstantakis.finance_app.data.models
 
 data class ProfitInfoDomain(
     var id: Int,
-    var profit: Double
+    var profit: Int
 ) {
 }
